+++ sdl3-sys/ChangeLog.md
@@ -1,5 +1,8 @@
 # Changes
 
+- 0.5.5:
+    - Update SDL to 3.2.22
+
 - 0.5.4:
     - Update SDL to 3.2.20